feat: CircuitBreaker

+++ scrapper/src/main/java/backend/academy/scrapper/controller/FilterController.java
@@ -6,10 +6,7 @@ import backend.academy.scrapper.dto.response.filter.FilterResponse;
 import backend.academy.scrapper.service.AccessFilterService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.HttpMethod;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.HttpStatusCode;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -18,10 +15,8 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.client.HttpClientErrorException;
 import org.springframework.web.client.HttpServerErrorException;
-import org.springframework.web.reactive.function.client.WebClientRequestException;
-import org.springframework.web.reactive.function.client.WebClientResponseException;
-import java.net.URI;
 
 @RestController
 @RequestMapping("/filter")
@@ -34,20 +29,17 @@ public class FilterController {
     @PostMapping("/{tgChatId}")
     @ResponseStatus(HttpStatus.CREATED)
     public FilterResponse createFilter(@PathVariable Long tgChatId, @RequestBody FilterRequest filterRequest) {
-        log.info("POST /filter/{tgChatId}/create");
+        log.info("POST /filter/{tgChatId}");
 
+        //        throw new HttpServerErrorException(
+        //            HttpStatus.INTERNAL_SERVER_ERROR,
+        //            "Сервер сломался по-настоящему"
+        //        );
 
-        // Имитация ошибок для тестирования (10% вероятность ошибки)
-
-//        throw new WebClientResponseException(
-//            503,
-//            "Service Unavailable",
-//            HttpHeaders.EMPTY,
-//            null,
-//            null
-//        );
-
-
+        //        throw new HttpClientErrorException(
+        //            HttpStatus.NOT_FOUND,
+        //            "Сервер сломался по-настоящему"
+        //        );
         return accessFilterService.createFilter(tgChatId, filterRequest);
     }
 
@@ -62,6 +54,21 @@ public class FilterController {
     @ResponseStatus(HttpStatus.OK)
     public FilterResponse deleteFilter(@PathVariable Long tgChatId, @RequestBody FilterRequest filterRequest) {
         log.info("DELETE /filter/{tgChatId}/delete");
+        // 70% вероятность исключения
+        if (Math.random() < 0.5) {
+            log.info("INTERNAL_SERVER_ERROR");
+
+            throw new HttpServerErrorException(
+                    HttpStatus.INTERNAL_SERVER_ERROR, "Серверная ошибка (тестовая, 70% вероятность)");
+        }
+
+        if (Math.random() < 0.5) {
+            log.info("NOT_FOUND");
+
+            throw new HttpClientErrorException(HttpStatus.NOT_FOUND, "Сервер сломался по-настоящему");
+        }
+        log.info("ResponseException");
+
         return accessFilterService.deleteFilter(tgChatId, filterRequest);
     }
 }
